+++ angelcloud-provider/angelcloud-provider-blog/src/main/java/com/angel/provider/model/domain/BlogArticle.java
@@ -77,7 +77,7 @@ public class BlogArticle extends BaseEntity implements Serializable {
     /**
      * 浏览量
      */
-    private Integer browseCount;
+    private Long browseCount;
 
     /**
      * 外链
